requirements.txt
--index-url https://download.pytorch.org/whl/cu121
flask==3.0.2
flask-dropzone==1.6.0
openai-whisper==20231117
torch==2.7.0
torchaudio==2.7.0
numpy==1.26.4
psutil==5.9.8
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.2
ffmpeg-python==0.2.0
transformers>=4.30.0 